Changed all doubles to floats in IonWake_105 GPU devices are mainly single precision -- doubles caused some kernel launch failures

+++ source/IonWake_105_ionColl.h
@@ -227,7 +227,7 @@ __global__ void setCollisionList_105
 *
 */
 
-__device__ double errorFn_inv(double y); 
+__device__ float errorFn_inv(float y); 
 
 //--------------------------------------------------------------------
 // sampling of Maxwellian distributions :
